Refactor of translator classes, renamed utils.py to sheet.py
